Moved unit tests from test_integration to test_unit
